% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster.R
\name{cluster}
\alias{cluster}
\title{cluster}
\usage{
cluster(numberOfColumns)
}
\arguments{
\item{numberOfColumns}{Number of columns of the distanceMatrixDF data frame. The input will be automatically cenerated by the clusterComputation function.}
}
\description{
cluster is a function to find the minimum value of each column.
It creates an empty vector with the same length as the classMeteroites variable.
It slice the distanceMatrixDF data frame into single columnns and comparing there every row with the minimum value.
If the function find the minimum value it saves the position of the row into the emptyClassMeteroites vector.
It returns the emptyClassMeteroites vector

WARNING:
Just works in combination with the clusterComputation function.
}
