% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster.R
\name{clusterComputation}
\alias{clusterComputation}
\title{clusterComputation}
\usage{
clusterComputation()
}
\value{
Returns a data frame with the classes and the number of meteroites in each class.
}
\description{
The \code{clusterComputation} function cluster the different subtypes of meteroites to the related main types together.
The function contains a vector with all the main types of meteroites.
The matrix from the \code{\link{distance}} function will be sliced into single columns.
In each column the index position of the smalles value will be calculated.
For each minimal value in an empty vector with the same length as the vector for the main types +1 value is added.
The function returns a data frame made out of the main type of meteroites and the empty vector with the added values at the index position with the smalles value.
The function uses \code{sapply} to loop the computation.
}
\seealso{
\code{\link[base]{sapply}},
\code{\link{distance}}
}
