% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competition.R
\name{competition}
\alias{competition}
\title{competition}
\usage{
competition()
}
\description{
A function to demonstarte the advantage of parallel computing on big data sets.
For a better demonstrating the distance function result is coppied several times into a data frame to increase the data volume.
}
