% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distance.R
\name{distance}
\alias{distance}
\title{distance}
\usage{
distance()
}
\value{
Returns a matrix with the string distances between main and subtypes of meteroites.
}
\description{
The distance function computes the string distances between main and subtypes of meteroites.
For measuring the distance between to strings the Jaro-Winker distance is used.
A value of 1 means there is no similarity between two strings.
A value of 0 means the similarity between two strings is 100%.
This means that a low value indicates a possible similarity, but this does not mean that two objects are from the same main type.
}
