% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{meteroites.hist}
\alias{meteroites.hist}
\title{meteroites.hist}
\usage{
meteroites.hist(l)
}
\arguments{
\item{l}{defines the x limit of the included mass in the histogram. The limit starts from 0.}
}
\value{
A histogram with the included mass between 0 and l.
}
\description{
\code{meteroites.hist} is a function for plotting a histogram of the meteroits depending on the mass.
}
\examples{
\dontrun{meteroites.hist(10000)}
}
