% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{meteroites.max}
\alias{meteroites.max}
\title{meteroites.max}
\usage{
meteroites.max()
}
\arguments{
\item{No}{additional argument is used}
}
\description{
\code{meteroites.max} is a function for finding the maximum of the mass of the meteroites.
Prints an output in the console.
}
