% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{meteroites.mean}
\alias{meteroites.mean}
\title{meteroites.mean}
\usage{
meteroites.mean()
}
\arguments{
\item{No}{additional argument is used}
}
\description{
\code{meteroites.mean} is a function for finding the mean of the mass of the meteroites.
Prints an output in the consol.
}
