% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{meteroites.min}
\alias{meteroites.min}
\title{meteroites.min}
\usage{
meteroites.min()
}
\arguments{
\item{No}{additional argument is used}
}
\description{
\code{meteroites.min} is a function for finding the minimum of the mass of the meteroites.
Prints an output in the console.
}
