% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{meteroites.world}
\alias{meteroites.world}
\title{meteroites.world}
\usage{
meteroites.world()
}
\value{
Returns a graph of the world with the locations of founded meteroites.
}
\description{
\code{meteroites.world} is a function for plotting a world map with the location of each meteroites from the meteroitesData data frame.
}
