% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{meteroitesGUI}
\alias{meteroitesGUI}
\title{meteroitesGUI}
\usage{
meteroitesGUI()
}
\value{
Returns a shiny app to analyse and visualize the meteroites data.
}
\description{
A function to run the meteroites2 shiny application. It allows to analyse and visualize the meteroites data.
The shiny application contains the \code{\link{clusterComputation}}, \code{\link{meteroites.plotClass}},
\code{\link{meteroites.heatMap}}, \code{\link{meteroites.world}} and \code{\link{meteroites.hist}} function.
}
